+++ src/mlpack/methods/reinforcement_learning/td3_impl.hpp
@@ -1,21 +1,21 @@
 /**
- * @file methods/reinforcement_learning/ddpg_impl.hpp
+ * @file methods/reinforcement_learning/td3_impl.hpp
  * @author Tarek Elsayed
  *
- * This file is the implementation of DDPG class, which implements the
- * Deep Deterministic Policy Gradient algorithm.
+ * This file is the implementation of TD3 class, which implements the
+ * Twin Delayed Deep Deterministic Policy Gradient algorithm.
  *
  * mlpack is free software; you may redistribute it and/or modify it under the
  * terms of the 3-clause BSD license.  You should have received a copy of the
  * 3-clause BSD license along with mlpack.  If not, see
  * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
-#ifndef MLPACK_METHODS_RL_DDPG_IMPL_HPP
-#define MLPACK_METHODS_RL_DDPG_IMPL_HPP
+#ifndef MLPACK_METHODS_RL_td3_IMPL_HPP
+#define MLPACK_METHODS_RL_td3_IMPL_HPP
 
 #include <mlpack/prereqs.hpp>
 
-#include "ddpg.hpp"
+#include "td3.hpp"
 
 namespace mlpack {
 
@@ -23,29 +23,25 @@ template <
   typename EnvironmentType,
   typename QNetworkType,
   typename PolicyNetworkType,
-  typename NoiseType,
   typename UpdaterType,
   typename ReplayType
 >
-DDPG<
+TD3<
   EnvironmentType,
   QNetworkType,
   PolicyNetworkType,
-  NoiseType, 
   UpdaterType,
   ReplayType
->::DDPG(TrainingConfig& config,
-       QNetworkType& learningQNetwork,
+>::TD3(TrainingConfig& config,
+       QNetworkType& learningQ1Network,
        PolicyNetworkType& policyNetwork,
-       NoiseType& noise,
        ReplayType& replayMethod,
        UpdaterType qNetworkUpdater,
        UpdaterType policyNetworkUpdater,
        EnvironmentType environment):
   config(config),
-  learningQNetwork(learningQNetwork),
+  learningQ1Network(learningQ1Network),
   policyNetwork(policyNetwork),
-  noise(noise),
   replayMethod(replayMethod),
   qNetworkUpdater(std::move(qNetworkUpdater)),
   #if ENS_VERSION_MAJOR >= 2
@@ -58,13 +54,12 @@ DDPG<
   environment(std::move(environment)),
   totalSteps(0),
   deterministic(false)
-{
-  // Reset the noise instance.
-  noise.reset();
-
+{ 
   // Set up q-learning and policy networks.
   targetPNetwork = policyNetwork;
-  targetQNetwork = learningQNetwork;
+  targetQ1Network = learningQ1Network;
+  learningQ2Network = learningQ1Network;
+  targetQ2Network = learningQ2Network;
 
   // Reset all the networks.
   // Note: the q and policy networks have an if condition before reset.
@@ -79,19 +74,22 @@ DDPG<
   const size_t networkSize = envSampleSize +
       policyNetwork.Network()[policyNetwork.Network().size() - 1]->OutputSize();
 
-  if (learningQNetwork.Parameters().n_elem != networkSize)
-    learningQNetwork.Reset(networkSize);
-  
-  targetQNetwork.Reset(networkSize);
+  if (learningQ1Network.Parameters().n_elem != networkSize)
+  {
+    learningQ1Network.Reset(networkSize);
+    learningQ2Network.Reset(networkSize);
+  }
+  targetQ1Network.Reset(networkSize);
+  targetQ2Network.Reset(networkSize);
 
   #if ENS_VERSION_MAJOR == 1
-  this->qNetworkUpdater.Initialize(learningQNetwork.Parameters().n_rows,
-                                   learningQNetwork.Parameters().n_cols);
+  this->qNetworkUpdater.Initialize(learningQ1Network.Parameters().n_rows,
+                                   learningQ1Network.Parameters().n_cols);
   #else
   this->qNetworkUpdatePolicy = new typename UpdaterType::template
       Policy<arma::mat, arma::mat>(this->qNetworkUpdater,
-                                   learningQNetwork.Parameters().n_rows,
-                                   learningQNetwork.Parameters().n_cols);
+                                   learningQ1Network.Parameters().n_rows,
+                                   learningQ1Network.Parameters().n_cols);
   #endif
 
   #if ENS_VERSION_MAJOR == 1
@@ -105,7 +103,8 @@ DDPG<
   #endif
 
   // Copy over the learning networks to their respective target networks.
-  targetQNetwork.Parameters() = learningQNetwork.Parameters();
+  targetQ1Network.Parameters() = learningQ1Network.Parameters();
+  targetQ2Network.Parameters() = learningQ2Network.Parameters();
   targetPNetwork.Parameters() = policyNetwork.Parameters();
 }
 
@@ -113,18 +112,16 @@ template <
   typename EnvironmentType,
   typename QNetworkType,
   typename PolicyNetworkType,
-  typename NoiseType,
   typename UpdaterType,
   typename ReplayType
 >
-DDPG<
+TD3<
   EnvironmentType,
   QNetworkType,
   PolicyNetworkType,
-  NoiseType,
   UpdaterType,
   ReplayType
->::~DDPG()
+>::~TD3()
 {
   #if ENS_VERSION_MAJOR >= 2
   delete qNetworkUpdatePolicy;
@@ -136,21 +133,21 @@ template <
   typename EnvironmentType,
   typename QNetworkType,
   typename PolicyNetworkType,
-  typename NoiseType,
   typename UpdaterType,
   typename ReplayType
 >
-void DDPG<
+void TD3<
   EnvironmentType,
   QNetworkType,
   PolicyNetworkType,
-  NoiseType,
   UpdaterType,
   ReplayType
 >::SoftUpdate(double rho)
 {
-  targetQNetwork.Parameters() = (1 - rho) * targetQNetwork.Parameters() +
-      rho * learningQNetwork.Parameters();
+  targetQ1Network.Parameters() = (1 - rho) * targetQ1Network.Parameters() +
+      rho * learningQ1Network.Parameters();
+  targetQ2Network.Parameters() = (1 - rho) * targetQ2Network.Parameters() +
+      rho * learningQ2Network.Parameters();
   targetPNetwork.Parameters() = (1 - rho) * targetPNetwork.Parameters() +
       rho * policyNetwork.Parameters();
 }
@@ -159,15 +156,13 @@ template <
   typename EnvironmentType,
   typename QNetworkType,
   typename PolicyNetworkType,
-  typename NoiseType,
   typename UpdaterType,
   typename ReplayType
 >
-void DDPG<
+void TD3<
   EnvironmentType,
   QNetworkType,
   PolicyNetworkType,
-  NoiseType,
   UpdaterType,
   ReplayType
 >::Update()
@@ -188,12 +183,14 @@ void DDPG<
   arma::mat nextStateActions;
   targetPNetwork.Predict(sampledNextStates, nextStateActions);
 
+  // Compute the estimated next Q-values using the target Q-networks.
   arma::mat targetQInput = arma::join_vert(nextStateActions,
       sampledNextStates);
-  arma::rowvec Q;
-  targetQNetwork.Predict(targetQInput, Q);
-  arma::rowvec nextQ = sampledRewards + config.Discount() 
-      * ((1 - isTerminal) % Q);
+  arma::rowvec Q1, Q2;
+  targetQ1Network.Predict(targetQInput, Q1);
+  targetQ2Network.Predict(targetQInput, Q2);
+  arma::rowvec nextQ = sampledRewards + config.Discount() * ((1 - isTerminal)
+      % arma::min(Q1, Q2));
 
   arma::mat sampledActionValues(action.size, sampledActions.size());
   for (size_t i = 0; i < sampledActions.size(); i++)
@@ -201,82 +198,95 @@ void DDPG<
                                  (sampledActions[i].action);
   arma::mat learningQInput = arma::join_vert(sampledActionValues,
       sampledStates);
-  learningQNetwork.Forward(learningQInput, Q);
+  learningQ1Network.Forward(learningQInput, Q1);
+  learningQ2Network.Forward(learningQInput, Q2);
 
-  arma::mat gradQLoss;
-  lossFunction.Backward(Q, nextQ, gradQLoss);
+  arma::mat gradQ1Loss, gradQ2Loss;
+  lossFunction.Backward(Q1, nextQ, gradQ1Loss);
+  lossFunction.Backward(Q2, nextQ, gradQ2Loss);
 
-  // Update the critic network.
-  arma::mat gradientQ;
-  learningQNetwork.Backward(learningQInput, gradQLoss, gradientQ);
+  // Sum both losses
+  arma::mat combinedLoss = gradQ1Loss + gradQ2Loss;
+
+  // Update the critic networks.
+  arma::mat gradientQ1, gradientQ2;
+  learningQ1Network.Backward(learningQInput, combinedLoss, gradientQ1);
+  learningQ2Network.Backward(learningQInput, combinedLoss, gradientQ2);
   #if ENS_VERSION_MAJOR == 1
-  qNetworkUpdater.Update(learningQNetwork.Parameters(), config.StepSize(),
-      gradientQ);
+  qNetworkUpdater.Update(learningQ1Network.Parameters(), config.StepSize(),
+      gradientQ1);
   #else
-  qNetworkUpdatePolicy->Update(learningQNetwork.Parameters(),
-      config.StepSize(), gradientQ);
+  qNetworkUpdatePolicy->Update(learningQ1Network.Parameters(),
+      config.StepSize(), gradientQ1);
+  #endif
+  #if ENS_VERSION_MAJOR == 1
+  qNetworkUpdater.Update(learningQ2Network.Parameters(), config.StepSize(),
+      gradientQ2);
+  #else
+  qNetworkUpdatePolicy->Update(learningQ2Network.Parameters(),
+      config.StepSize(), gradientQ2);
   #endif
 
   // Actor network update.
 
-  // Get the size of the first hidden layer in the Q network.
-  size_t hidden1 = learningQNetwork.Network()[0]->OutputSize();
-
-  arma::mat gradient;
-  for (size_t i = 0; i < sampledStates.n_cols; i++)
-  {
-    arma::mat grad, gradQ, q;
-    arma::colvec singleState = sampledStates.col(i);
-    arma::colvec singlePi;
-    policyNetwork.Forward(singleState, singlePi);
-    arma::colvec input = arma::join_vert(singlePi, singleState);
-    arma::mat weightLastLayer;
-
-    // Note that we can use an empty matrix for the backwards pass, since the
-    // networks use EmptyLoss.
-    learningQNetwork.Forward(input, q);
-    learningQNetwork.Backward(input, arma::mat("-1"), gradQ);
-    weightLastLayer = arma::reshape(learningQNetwork.Parameters().
-          rows(0, hidden1 * singlePi.n_rows - 1), hidden1, singlePi.n_rows);
-
-    arma::colvec gradQBias = gradQ(input.n_rows * hidden1, 0,
-        arma::size(hidden1, 1));
-    arma::mat gradPolicy = weightLastLayer.t() * gradQBias;
-    policyNetwork.Backward(singleState, gradPolicy, grad);
-    if (i == 0)
-    {
-      gradient.copy_size(grad);
-      gradient.fill(0.0);
-    }
-    gradient += grad;
-  }
-  gradient /= sampledStates.n_cols;
-
-  #if ENS_VERSION_MAJOR == 1
-  policyUpdater.Update(policyNetwork.Parameters(), config.StepSize(), gradient);
-  #else
-  policyNetworkUpdatePolicy->Update(policyNetwork.Parameters(),
-      config.StepSize(), gradient);
-  #endif
-
-  // Update target networks
   if (totalSteps % config.TargetNetworkSyncInterval() == 0)
+  {
+    // Get the size of the first hidden layer in the Q network.
+    size_t hidden1 = learningQ1Network.Network()[0]->OutputSize();
+
+    arma::mat gradient;
+    for (size_t i = 0; i < sampledStates.n_cols; i++)
+    {
+        arma::mat grad, gradQ, q;
+        arma::colvec singleState = sampledStates.col(i);
+        arma::colvec singlePi;
+        policyNetwork.Forward(singleState, singlePi);
+        arma::colvec input = arma::join_vert(singlePi, singleState);
+        arma::mat weightLastLayer;
+
+        // Note that we can use an empty matrix for the backwards pass, since the
+        // networks use EmptyLoss.
+        learningQ1Network.Forward(input, q);
+        learningQ1Network.Backward(input, arma::mat("-1"), gradQ);
+        weightLastLayer = arma::reshape(learningQ1Network.Parameters().
+            rows(0, hidden1 * singlePi.n_rows - 1), hidden1, singlePi.n_rows);
+
+        arma::colvec gradQBias = gradQ(input.n_rows * hidden1, 0,
+            arma::size(hidden1, 1));
+        arma::mat gradPolicy = weightLastLayer.t() * gradQBias;
+        policyNetwork.Backward(singleState, gradPolicy, grad);
+        if (i == 0)
+        {
+            gradient.copy_size(grad);
+            gradient.fill(0.0);
+        }
+        gradient += grad;
+    }
+    gradient /= sampledStates.n_cols;
+
+    #if ENS_VERSION_MAJOR == 1
+    policyUpdater.Update(policyNetwork.Parameters(), config.StepSize(), gradient);
+    #else
+    policyNetworkUpdatePolicy->Update(policyNetwork.Parameters(),
+        config.StepSize(), gradient);
+    #endif
+
+    // Update target networks
     SoftUpdate(config.Rho());
+  }
 }
 
 template <
   typename EnvironmentType,
   typename QNetworkType,
   typename PolicyNetworkType,
-  typename NoiseType,
   typename UpdaterType,
   typename ReplayType
 >
-void DDPG<
+void TD3<
   EnvironmentType,
   QNetworkType,
   PolicyNetworkType,
-  NoiseType,
   UpdaterType,
   ReplayType
 >::SelectAction()
@@ -287,9 +297,9 @@ void DDPG<
 
   if (!deterministic)
   {
-    arma::colvec sample = noise.sample() * 0.1;
-    sample = arma::clamp(sample, -0.25, 0.25);
-    outputAction = outputAction + sample;
+    arma::colvec noise = arma::randn<arma::colvec>(outputAction.n_rows) * 0.1;
+    noise = arma::clamp(noise, -0.25, 0.25);
+    outputAction = outputAction + noise;
   }
   action.action = arma::conv_to<std::vector<double>>::from(outputAction);
 }
@@ -298,15 +308,13 @@ template <
   typename EnvironmentType,
   typename QNetworkType,
   typename PolicyNetworkType,
-  typename NoiseType,
   typename UpdaterType,
   typename ReplayType
 >
-double DDPG<
+double TD3<
   EnvironmentType,
   QNetworkType,
   PolicyNetworkType,
-  NoiseType,
   UpdaterType,
   ReplayType
 >::Episode()